+++ CHANGELOG.md
@@ -6,6 +6,7 @@ versioning is done in a continuous fashion without worries of breaking changes.
 
 ## patches
 
+- `soft-serve`: host a git server having flavor for a terminal kind 2026-03-07
 - `tmux`: copies selections across ssh boundaries to paste whatever 2026-03-07
 - `claude`: alias work versions for the latest models provided kind 2026-03-03
 - `darwin`: increase mouse driver cursor pointer icon size for view 2026-03-01